Adding a smallestRepair flag to Mutation Result+bug fixes
This commit adds the implementation of a smallestRepair flag to the MutationResult. This flag indicates if a mutation is made to the immediate operator or operands of a Binary or Unary expression. This commit also fixes a few bugs with the ORO mutator and moves it into its own visitor